Refactor implementation using SIOCGIFFLAGS ioctl
Instead of spawning a background process in pre-up and communicating
with this process in up, only run the executor in the up phase and
check if the interface was up'ed prior to netlink socket creation
via the SIOCGIFFLAGS ioctl.

Now the code just starts a background thread using pthread and uses
a semaphore to communicate changes of the interface state.